(fix) dev packages only imported in test
Take out of common where they'd be imported in production deploy.
