+++ src/common/utils.py
@@ -1,8 +1,6 @@
 from django.shortcuts import _get_queryset
 from django.utils.http import urlencode
 from django.urls import reverse
-import numpy as np
-import cv2 as cv
 import qrcode
 
 
@@ -37,15 +35,3 @@ def get_object_or_none(klass, *args, **kwargs):
 
 def qr_image_from_string(s):
     return qrcode.make(s)
-
-
-def get_data_in_qr_image(img):
-    qr_detector = cv.QRCodeDetector()
-    data, bbox, _ = qr_detector.detectAndDecode(img)
-    if bbox is not None and len(bbox) > 0:
-        return data
-    return None
-
-
-def image_from_bytestring(bytestring):
-    return cv.imdecode(np.fromstring(bytestring, np.uint8), cv.IMREAD_COLOR)
